Initial commit
The application is a web scraping, image search and recognition tool for the Copywrite industry. The application scrapes images (and content within the image), and captions from specific websites. Tools such as NLP and CV are used extract metadata about the image, e.g., artwork within the image. The resulting output is used to find the name of the artist and or person who took the photo. The Copywrite industry can use this information make sure the artist is correctly attributed and compensated.